<!--
## Before you hit that Submit button....

Please read our [contribution guidelines](https://github.com/node-red/node-red-nodes/blob/master/CONTRIBUTING.md)
before submitting a pull-request.

## Types of changes

What types of changes does your code introduce?
Put an `x` in the boxes that apply
-->

- [x] Bugfix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)

<!--
If you want to raise a pull-request with a new feature, or a refactoring
of existing code, it **may well get rejected** if it hasn't been discussed on
the [forum](https://discourse.nodered.org) or
[slack team](https://nodered.org/slack) first.

-->

## Proposed changes

This PR is intended to increase the performance of Concurrent Request and to resolve some errors like CONNREST.
To do that by let MySQL Pool manage the connections instead of manually manage connections.

refer to: #804 

And FYI, there's some benchmark result below.
**Before optimize:** ab -n 500 -c 30
```
Document Path:          /data-flow/api/benchmark
Document Length:        13 bytes

Concurrency Level:      30
Time taken for tests:   13.528 seconds
Complete requests:      500
Failed requests:        0
Total transferred:      138500 bytes
HTML transferred:       6500 bytes
Requests per second:    36.96 [#/sec] (mean)
Time per request:       811.655 [ms] (mean)
Time per request:       27.055 [ms] (mean, across all concurrent requests)
Transfer rate:          10.00 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:       55  693 2318.4     88    9888
Processing:    59  102  34.5     93     204
Waiting:       59  102  34.5     93     203
Total:        117  795 2316.3    189    9995

Percentage of the requests served within a certain time (ms)
  50%    189
  66%    210
  75%    239
  80%    271
  90%    500
  95%   9902
  98%   9994
  99%   9994
 100%   9995 (longest request)
```

**After optimize:** ab -n 500 -c 30
```
Document Path:          /data-flow/api/benchmark
Document Length:        13 bytes

Concurrency Level:      30
Time taken for tests:   0.477 seconds
Complete requests:      500
Failed requests:        0
Total transferred:      138500 bytes
HTML transferred:       6500 bytes
Requests per second:    1048.70 [#/sec] (mean)
Time per request:       28.607 [ms] (mean)
Time per request:       0.954 [ms] (mean, across all concurrent requests)
Transfer rate:          283.68 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        8   12   1.6     12      17
Processing:     9   14   2.4     14      21
Waiting:        9   14   2.4     14      21
Total:         19   26   2.8     26      34

Percentage of the requests served within a certain time (ms)
  50%     26
  66%     27
  75%     28
  80%     29
  90%     30
  95%     32
  98%     32
  99%     33
 100%     34 (longest request)
```

## Checklist
<!-- Put an `x` in the boxes that apply -->

- [x] I have read the [contribution guidelines](https://github.com/node-red/node-red-nodes/blob/master/CONTRIBUTING.md)
- [ ] For non-bugfix PRs, I have discussed this change on the forum/slack team.
- [x] I have run `grunt` to verify the unit tests pass
- [ ] I have added suitable unit tests to cover the new/changed functionality