@dceejay 2021-06-25
Hi @ericdum 
any chance you can have a look at this please ? https://github.com/node-red/node-red-nodes/issues/811
seems to have occurred since we merged your PR. 
Thanks